Reimplementation of Colorset
The final version of the project from 2019 seems to have been altered at some point, perhaps due to a commit from an earlier branch. This fixes the color scheme to once again using a gradient set. Perhaps look into why the backend loading is not as efficient as it used to be.